Relocate to spam.ts
