+++ app/core/RPCMethods/RPCMethodMiddleware.test.ts
@@ -32,8 +32,7 @@ import {
 } from '../Permissions/specifications';
 import { EthAccountType, EthMethod } from '@metamask/keyring-api';
 
-jest.mock('./utils', () => ({
-  ...jest.requireActual('./utils'),
+jest.mock('./spam', () => ({
   processOriginThrottlingRejection: jest.fn(),
   validateOriginThrottling: jest.fn(),
 }));
